feat: 강한/상충되는 기운 필드 추가
