+++ src/four_pillars/entities/enums.py
@@ -1,0 +1,10 @@
+from enum import Enum
+
+
+class FiveElements(str, Enum):
+    """오행 (五行)"""
+    WOOD = "목(木)"  # 목
+    FIRE = "화(火)"  # 화
+    EARTH = "토(土)"  # 토
+    METAL = "금(金)"  # 금
+    WATER = "수(水)"  # 수
