+++ src/four_pillars/entities/schemas.py
@@ -1,15 +1,5 @@
-from enum import Enum
 from typing import Optional, List
 from typing_extensions import TypedDict
-
-
-class FiveElements(str, Enum):
-    """오행 (五行)"""
-    WOOD = "목(木)"  # 목
-    FIRE = "화(火)"  # 화
-    EARTH = "토(土)"  # 토
-    METAL = "금(金)"  # 금
-    WATER = "수(水)"  # 수
 
 
 class FourPillar(TypedDict, total=False):
