+++ src/users/router.py
@@ -65,5 +65,7 @@ async def get_lotto_recommendation(
     - num5, num6: 최근 자주 나온 번호 (1-45)
     - cold_nums: 기운과 상충하는 숫자 (1-3개)
     - infrequent_nums: 등장 빈도가 낮은 숫자 (1-3개)
+    - strong_element: 강한 기운
+    - weak_element: 상충되는 기운
     """
     return await lotto_service.get_lotto_recommendation(user_id=user_id)
